+++ Sources/RichTextKit/Actions/RichTextAction+KeyboardShortcutModifier.swift
@@ -1,5 +1,5 @@
 //
-//  View+RichTextAction.swift
+//  RichTextAction+KeyboardShortcutModifier.swift
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-12-13.
@@ -11,13 +11,11 @@ import SwiftUI
 public extension RichTextAction {
  
     /**
-     This view modifier can be used to apply keyboard action
-     shortcuts to any view.
+     This view modifier can apply keyboard shortcuts for any
+     ``RichTextAction`` to any view.
      
      You can also apply it with the `.keyboardShortcut(for:)`
      view extension.
-     
-     > Note: This modifier only has effect on some platform.
      */
     struct KeyboardShortcutModifier: ViewModifier {
         
@@ -28,31 +26,31 @@ public extension RichTextAction {
         private let action: RichTextAction
         
         public func body(content: Content) -> some View {
-            #if iOS || macOS || os(visionOS)
-            switch action {
-            case .copy: content.keyboardShortcut("c", modifiers: .command)
-            case .dismissKeyboard: content
-            case .print: content.keyboardShortcut("p", modifiers: .command)
-            case .redoLatestChange: content.keyboardShortcut("z", modifiers: [.command, .shift])
-            case .setAlignment(let align): content.keyboardShortcut(for: align)
-            case .stepFontSize(let points): content.keyboardShortcut(points < 0 ? "-" : "+", modifiers: .command)
-            case .stepIndent(let steps): content.keyboardShortcut(steps < 0 ? "Ö" : "Ä", modifiers: .command)
-            case .stepSuperscript: content
-            case .toggleStyle(let style): content.keyboardShortcut(for: style)
-            case .undoLatestChange: content.keyboardShortcut("z", modifiers: .command)
-            }
-            #else
-            content
-            #endif
+            content.keyboardShortcut(for: action)
         }
     }
 }
 
 public extension View {
 
-    /// Apply a rich text action keyboard shortcut.
+    /// Apply a ``RichTextAction/KeyboardShortcutModifier``.
     @ViewBuilder
     func keyboardShortcut(for action: RichTextAction) -> some View {
-        modifier(RichTextAction.KeyboardShortcutModifier(action))
+        #if iOS || macOS || os(visionOS)
+        switch action {
+        case .copy: keyboardShortcut("c", modifiers: .command)
+        case .dismissKeyboard: self
+        case .print: keyboardShortcut("p", modifiers: .command)
+        case .redoLatestChange: keyboardShortcut("z", modifiers: [.command, .shift])
+        case .setAlignment(let align): keyboardShortcut(for: align)
+        case .stepFontSize(let points): keyboardShortcut(points < 0 ? "-" : "+", modifiers: .command)
+        case .stepIndent(let steps): keyboardShortcut(steps < 0 ? "Ö" : "Ä", modifiers: .command)
+        case .stepSuperscript: self
+        case .toggleStyle(let style): keyboardShortcut(for: style)
+        case .undoLatestChange: keyboardShortcut("z", modifiers: .command)
+        }
+        #else
+        self
+        #endif
     }
 }