+++ Sources/RichTextKit/Alignment/RichTextAlignment+KeyboardShortcutModifier.swift
@@ -1,5 +1,5 @@
 //
-//  View+RichTextAlignment.swift
+//  RichTextAlignment+KeyboardShortcutModifier.swift
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2023-10-17.
@@ -7,6 +7,29 @@
 //
 
 import SwiftUI
+
+public extension RichTextAlignment {
+ 
+    /**
+     This view modifier can apply keyboard shortcuts for any
+     ``RichTextAlignment`` to any view.
+     
+     You can also apply it with the `.keyboardShortcut(for:)`
+     view extension.
+     */
+    struct KeyboardShortcutModifier: ViewModifier {
+        
+        public init(_ alignment: RichTextAlignment) {
+            self.alignment = alignment
+        }
+        
+        private let alignment: RichTextAlignment
+        
+        public func body(content: Content) -> some View {
+            content.keyboardShortcut(for: alignment)
+        }
+    }
+}
 
 public extension View {