+++ Sources/RichTextKit/_Deprecated/RichTextAlignment+Deprecated.swift
@@ -1,0 +1,4 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "RichTextAlignment.Picker")
+public typealias RichTextAlignmentPicker = RichTextAlignment.Picker